removed isConsecutiveDays.js and moved its function to incrementStreak.js. modified auditLog schema for loginId to refer to users collection
