+++ models/auditLog.js
@@ -1,13 +1,13 @@
 const mongoose = require('mongoose')
 const { Schema } = mongoose
 
-const AuditLogSchema = new Schema({
-    loginId: {type: String, required: true, unique: true},
+const AuditLogsSchema = new Schema({
+    loginId: {type: String, required: true, unique: true, ref: 'users'},
     type: {type: String, default: null, enum: ['flashcard']},
+    shareId: {type: Number, required: true, unique: true, ref: 'collections',},
     timestamp: {type: Date, default: null},
-    shareId: {type: Number, required: true, unique: true}
 }, {
     timestamps: true,
 })
 
-module.exports = mongoose.model('auditlogs', AuditLogSchema)
+module.exports = mongoose.model('AuditLogs', AuditLogsSchema)
